solidpython_testing/requirements.txt
euclid3==0.1
pkg-resources==0.0.0
prettytable==0.7.2
pypng==0.0.19
regex==2019.12.20
solidpython==0.4.6
